Unlimited Instances of a Component, with Independent States & Shared States
For this one there where some ways I've considered:

In my Redux store I render unlimited (state-based) number of instances of the same sub-app component. In the current situation, there is one store for all of the instances, and therefore a change in one component happened relatively on the others. There are some states that I would like to keep working in the same current way, but I would like the others to be independent for each instance of the component, in such a way that a change of state wouldn't affect the same state on other instances.

One solution I've tried was to define a state with hook for the independent-needed states. That worked well for a state I've used in one container, but I've got "too many renders" error for the states I've used in more than one container.

Another solution I've been thinking about is to create a store for the whole app, and a sub-store for every instance of the component, but I wonder how will a container know to which store does it need to call.

What I decided was to put parameter "instance id" on the action and therefore keep working with one main store for the app.