Click on Non-Current Month's Day, Will Change to Day's Month, and Near Calendar Update Relatively
I had the following problem:

When the user clicks on a non-current month's day, I'd like to change the month to the day's month. But what if the changed month is the same of the near calendar's month? I'd like to change it's month too. The problem was that when I clicked on left calendar to go to next month, the right calendar saw that they were equal and increased month, while the same left calendar saw that too, and decreased. Solution was to add parameter "lastChangedId" to Redux's store, and that way be notified of which calendar updated last. The solution raised the error :index.js:1 Warning: Cannot update a component (`ConnectFunction`) while rendering a different component (`Header`), that needs to be solved too.