Fixed DayElement Bug & Created Months Change on Select Bug
The bug is pretty unexplained to me. I changed the isInRange and isSelected parameters to be calculated and based on parameters, instead of being states. That probably what made them to reset on every render as I wished. I don't quite sure why it has to happened when dayElement got called from a Redux's container (while from a react component state was the right choise).