Added Date Disabling Feature. Added Left Board Non-Current Dates Click that Calls Right Board to be Upper by 1, and so on the Right to the Left.
Date Disabling Feature:
I demand the user to send me a dates (or nothing beacause of defaults) from the type of the js obect date (I throw errors). Then year selector, month selector and the calendar will disable values relatively.

Added Left Board Non-Current Dates Click that Calls Right Board to be Upper by 1, and so on the Right to the Left:
From now on, a click on the left to the non current dates that are on the left, while the right board renders a month that is not the +1 of the left. The left will change it to its +1.